--- components/fhir-core/src/main/java/org/wso2/healthcare/codegen/tool/framework/fhir/core/versions/r5/model/FHIRR5ImplementationGuide.java
@@ -1,4 +1,0 @@
-package org.wso2.healthcare.codegen.tool.framework.fhir.core.versions.r5.model;
-
-public class FHIRR5ImplementationGuide {
-}
